Change multimodal to one client
