+++ adalflow/adalflow/components/model_client/__init__.py
@@ -64,10 +64,6 @@ get_probabilities = LazyImport(
     "adalflow.components.model_client.openai_client.get_probabilities",
     OptionalPackages.OPENAI,
 )
-OpenAIMultimodalClient = LazyImport(
-    "adalflow.components.model_client.openai_multimodal_client.OpenAIMultimodalClient",
-    OptionalPackages.OPENAI,
-)
 
 __all__ = [
     "CohereAPIClient",
@@ -80,7 +76,6 @@ __all__ = [
     "GroqAPIClient",
     "OpenAIClient",
     "GoogleGenAIClient",
-    "OpenAIMultimodalClient",
 ]
 
 for name in __all__:
